requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
jinja2==3.1.4
apscheduler==3.10.4
python-multipart==0.0.9
